fix: 회원가입 - 프로필 사진 변경 ActionSheet, 이미지 오류 수정
