+++ 14th-team5-iOS/App/Sources/Presentation/Account/AccountSignUp/Reactor/AccountSignUpReactor.swift
@@ -30,7 +30,6 @@ public final class AccountSignUpReactor: Reactor {
         case setDay(Int?)
         case didTapDateNextButton
         
-        case profileImageTapped // Action Sheet출력 하는 이벤트
         case didTapCompletehButton
         case profilePresignedURL(String, Data)
         case didTapPHAssetsImage(Data)
@@ -46,7 +45,6 @@ public final class AccountSignUpReactor: Reactor {
         case didTapDateNextButton
         case setEditNickName(AccountNickNameEditResponse?)
         
-        case profileImageTapped
         case setprofilePresignedURL(String)
         case setprofileImage(Data)
         case didTapCompletehButton(AccessTokenResponse?)
@@ -72,7 +70,6 @@ public final class AccountSignUpReactor: Reactor {
         var profileType: AccountLoaction = .account
         
         var profilePresignedURL: String = ""
-        var profileImageButtontapped: Bool = false
         var profileImage: Data? = nil
         var didTapCompletehButtonFinish: AccessTokenResponse? = nil
     }
@@ -110,9 +107,6 @@ extension AccountSignUpReactor {
             return Observable.just(Mutation.didTapDateNextButton)
             
             // MARK: Profile
-        case .profileImageTapped:
-            return Observable.just(Mutation.profileImageTapped)
-        
         case let .profilePresignedURL(presignedURL, originImage):
             return .concat(
                 .just(.setprofilePresignedURL(presignedURL)),
@@ -202,8 +196,6 @@ extension AccountSignUpReactor {
             newState.dateButtonTappedFinish = true
         case .setprofilePresignedURL(let url):
             newState.profilePresignedURL = url
-        case .profileImageTapped:
-            newState.profileImageButtontapped = true
         case let .setprofileImage(profileImage):
             newState.profileImage = profileImage
         case .didTapCompletehButton(let token):
