+++ 14th-team5-iOS/App/Sources/Presentation/Account/AccountSignUp/ViewControllers/AccountSignUpViewController.swift
@@ -8,21 +8,11 @@
 import UIKit
 import Core
 import DesignSystem
-import PhotosUI
 
 fileprivate typealias _Str = AccountSignUpStrings
 public final class AccountSignUpViewController: BasePageViewController<AccountSignUpReactor> {
     private let nextButton = UIButton()
     private let descLabel = UILabel()
-    
-    private var pickerConfiguration: PHPickerConfiguration = {
-        var configuration: PHPickerConfiguration = PHPickerConfiguration()
-        configuration.filter = .images
-        configuration.selectionLimit = 1
-        configuration.selection = .default
-        return configuration
-    }()
-    private lazy var profilePickerController: PHPickerViewController = PHPickerViewController(configuration: pickerConfiguration)
     
     private var pages = [UIViewController]()
     private var initalPage = 0
@@ -31,7 +21,6 @@ public final class AccountSignUpViewController: BasePageViewController<AccountSi
         super.viewDidLoad()
         
         dataSource = self
-        profilePickerController.delegate = self
         setViewControllers([pages[initalPage]], direction: .forward, animated: true)
         isPagingEnabled = false
     }
@@ -50,24 +39,6 @@ public final class AccountSignUpViewController: BasePageViewController<AccountSi
             .withUnretained(self)
             .bind(onNext: { $0.0.goToNextPage() })
             .disposed(by: disposeBag)
-        
-        reactor.state.map { $0.profileImageButtontapped }
-            .filter { $0 }
-            .observe(on: Schedulers.main)
-            .withUnretained(self)
-            .bind(onNext: { $0.0.createAlertController(owner: $0.0) })
-            .disposed(by: disposeBag)
-        
-        NotificationCenter.default.rx
-            .notification(.PHPickerAssetsDidFinishPickingProcessingPhotoNotification)
-            .compactMap { notification -> Data? in
-                guard let userInfo = notification.userInfo else { return nil }
-                return userInfo["selectImage"] as? Data
-            }
-            .map{ Reactor.Action.didTapPHAssetsImage($0)}
-            .bind(to: reactor.action)
-            .disposed(by: disposeBag)
-        
     }
     
     public override func setupUI() {
@@ -109,45 +80,5 @@ extension AccountSignUpViewController: UIPageViewControllerDataSource {
         
         guard currentIndex < pages.count - 1 else { return nil }
         return pages[currentIndex + 1]
-    }
-}
-
-
-extension AccountSignUpViewController {
-    private func createAlertController(owner: AccountSignUpViewController) {
-        let alertController: UIAlertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
-        let presentCameraAction: UIAlertAction = UIAlertAction(title: "카메라", style: .default) { _ in
-            let cameraViewController = CameraDIContainer(cameraType: .profile).makeViewController()
-            owner.navigationController?.pushViewController(cameraViewController, animated: true)
-        }
-        let presentAlbumAction: UIAlertAction = UIAlertAction(title: "앨범", style: .default) { _ in
-            self.profilePickerController.modalPresentationStyle = .fullScreen
-            self.profilePickerController.overrideUserInterfaceStyle = .dark
-            self.present(self.profilePickerController, animated: true)
-        }
-        let presentDefaultAction: UIAlertAction = UIAlertAction(title: "초기화", style: .destructive) { _ in
-            print("초기화 구문")
-        }
-        let presentCancelAction: UIAlertAction = UIAlertAction(title: "취소", style: .cancel, handler: nil)
-        
-        [presentCameraAction, presentAlbumAction, presentDefaultAction, presentCancelAction].forEach {
-            alertController.addAction($0)
-        }
-        alertController.overrideUserInterfaceStyle = .dark
-        owner.present(alertController, animated: true)
-    }
-}
-
-extension AccountSignUpViewController: PHPickerViewControllerDelegate {
-    public func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
-        let itemProvider = results.first?.itemProvider
-        picker.dismiss(animated: true)
-        if let imageProvider = itemProvider, imageProvider.canLoadObject(ofClass: UIImage.self) {
-            imageProvider.loadObject(ofClass: UIImage.self) { image, error in
-                guard let photoImage: UIImage = image as? UIImage,
-                      let originalData: Data = photoImage.jpegData(compressionQuality: 1.0) else { return }
-                imageProvider.didSelectProfileImageWithProcessing(photo: originalData, error: error)
-            }
-        }
     }
 }
